Extract an App Container Component from TodoForm component that renders the TodoForm. #11
