+++ client/index.js
@@ -1,5 +1,5 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
-import TodoForm from './todo-form'
+import App from './app'
 
-ReactDOM.render(<TodoForm />, document.querySelector('#app'))
+ReactDOM.render(<App />, document.querySelector('#app'))
